Move logic to control if BlockMenu is open into a hook and into Tools component.
